feat: Add admin API for song management with bulk import functionality and a corresponding UI modal.

+++ src/pages/api/admin/songs.ts
@@ -52,7 +52,7 @@ export default async function handler(
             continue;
           }
 
-          const difficulty = s.difficulty ? (typeof s.difficulty === 'string' ? parseInt(s.difficulty) : s.difficulty) : null;
+          const difficulty = s.difficulty ? (typeof s.difficulty === 'string' ? parseFloat(s.difficulty) : s.difficulty) : null;
 
           const newSong = {
             title: s.title || "Unknown Title",
